pyproject.toml
[project]
name = "add-project-name-here"
version = "0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "bblocks-data-importers (>=0.4.1,<0.5.0)",
    "bblocks-places (>=0.0.3,<0.0.4)",
    "datacommons-client[pandas]>=2.1.1",
    "oda-data>=2.1.2",
    "pandas (>=2.3.2,<3.0.0)",
]

[dependency-groups]
dev = [
    "black (>=25.9.0,<26.0.0)",
    "ruff>=0.13.1",
]
